requirements.txt
django==2.2.7
django-filter==2.2.0
djangorestframework==3.10.3
# for apscheduler should less than 3.0
tzlocal>=2,<3
apscheduler==3.6.3
gunicorn==20.0.4
psycopg2==2.8.6
oss2
redis
django-redis
opencv-python
numpy
channels>=2,<3
channels-redis>=3,<4
daphne>=2,<3
systemd-python

# for LIS
requests